+++ property_tests/tests/service/test/test.ts
@@ -1,13 +1,12 @@
-import { Principal } from '@dfinity/principal';
-import { execSync } from 'child_process';
 import fc from 'fast-check';
 
 import { runPropTests } from 'azle/property_tests';
-import { CanisterArb } from 'azle/property_tests/arbitraries/canister_arb';
-import { CandidMeta } from 'azle/property_tests/arbitraries/candid/candid_arb';
 import { ServiceArb } from 'azle/property_tests/arbitraries/candid/reference/service_arb';
+import { CanisterArb } from 'azle/property_tests/arbitraries/canister_arb';
 import { QueryMethodArb } from 'azle/property_tests/arbitraries/query_method_arb';
-import { Test } from 'azle/test';
+
+import { generateBody } from './generate_body';
+import { generateTests } from './generate_tests';
 
 const UniqueServicesArray = fc.uniqueArray(ServiceArb, {
     selector: (entry) => entry.src.candidType
@@ -19,66 +18,3 @@ const AllServicesQueryMethod = QueryMethodArb(UniqueServicesArray, ServiceArb, {
 });
 
 runPropTests(CanisterArb(AllServicesQueryMethod));
-
-function generateBody(
-    paramNames: string[],
-    paramServices: CandidMeta<Principal>[],
-    returnService: CandidMeta<Principal>
-): string {
-    const paramsAreServices = paramServices
-        .map((param, index) => {
-            const paramName = `param${index}`;
-
-            const paramIsAService = `(${paramName} as any).principal.toText() === "${param.agentArgumentValue.toText()}"`;
-
-            const throwError = `throw new Error('${paramName} must be a Service');`;
-
-            return `if (!(${paramIsAService})) ${throwError}`;
-        })
-        .join('\n');
-
-    const returnStatement =
-        paramServices.length === 0 ? returnService.src.valueLiteral : `param0`;
-
-    return `
-        ${paramsAreServices}
-
-        return ${returnStatement};
-    `;
-}
-
-function generateTests(
-    functionName: string,
-    paramServices: CandidMeta<Principal>[],
-    returnService: CandidMeta<Principal>
-): Test[] {
-    return [
-        {
-            name: `service ${functionName}`,
-            test: async () => {
-                // Using execSync because the JS Agent has a bug expecting services
-                // to be ordered by hash or something.
-                // See https://forum.dfinity.org/t/topic/20885/14
-
-                const paramsString = paramServices
-                    .map(
-                        (service) =>
-                            `service "${service.agentArgumentValue.toText()}"`
-                    )
-                    .join();
-
-                const result = execSync(
-                    `dfx canister call canister ${functionName} '(${paramsString})'`
-                )
-                    .toString()
-                    .trim();
-
-                return {
-                    Ok:
-                        result ===
-                        `(service "${returnService.agentArgumentValue.toText()}")`
-                };
-            }
-        }
-    ];
-}
